+++ front/app/containers/Admin/pages-menu/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  pageHeader: {
+    id: 'app.containers.admin.pages-menu.pageHeader',
+    defaultMessage: 'Pages',
+  },
+});