+++ front/app/hooks/useNavbarItems.ts
@@ -9,20 +9,22 @@ import { truncateMultiloc } from 'utils/textUtils';
 
 export type TNavbarItemsState = INavbarItem[] | undefined | null | Error;
 
-export default function useNavbarItems() {
+export default function useNavbarItems(
+  { onlyDefault } = { onlyDefault: false }
+) {
   const [navbarItems, setNavbarItems] = useState<TNavbarItemsState>(undefined);
 
   useEffect(() => {
-    const subscription = navbarItemsStream().observable.subscribe(
-      (response) => {
-        isNilOrError(response)
-          ? setNavbarItems(response)
-          : setNavbarItems(truncateTitles(response.data));
-      }
-    );
+    const subscription = navbarItemsStream({
+      onlyDefault,
+    }).observable.subscribe((response) => {
+      isNilOrError(response)
+        ? setNavbarItems(response)
+        : setNavbarItems(truncateTitles(response.data));
+    });
 
     return () => subscription.unsubscribe();
-  }, []);
+  }, [onlyDefault]);
 
   return navbarItems;
 }